The most significant changes involve the introduction of a new `vm` (VectorMath) namespace, which contains several vector math functions. This has led to the removal of several functions from the `GlobalValues_PG.h` and `ShipMovementHandler_PG.h` files, such as `normalizeVector`, `distanceBetweenPoints`, `dot`, and `createVector`. These functions have been replaced with equivalent functions from the `vm` namespace in various files. The `move` method in `EnemyShipMovementHandler_PG.h` and `PlayerShipMovementHandler_PG.h` now takes an additional parameter of type `sf::Vector2f`. The `draw` method in `Circle_AG.h` and `Circle_AG.cpp` has been made a constant function.
List of changes:

1. The `draw` method in `Circle_AG.h` and `Circle_AG.cpp` has been made a constant function.
2. The `move` method in `EnemyShipMovementHandler_PG.h` and `PlayerShipMovementHandler_PG.h` now takes an additional parameter of type `sf::Vector2f`.
3. The `grid` method in `GlobalMap_PG.h` now takes an integer cast of `chunkSize.x` and `chunkSize.y`.
4. Several functions in `GlobalValues_PG.h` have been removed, including `normalizeVector`, `distanceBetweenPoints`, and `limitVector`.
5. The `createVector` method in `GlobalValues_PG.h` and `GlobalValues_PG.cpp` has been removed.
6. The `normalize` method in `GlobalWindController_PG.h` and `GlobalWindController_PG.cpp` has been replaced with `vm::normalize`.
7. The `setWindDirection` method in `GlobalWindController_PG.h` now uses `vm::normalize`.
8. The `randomizeWind` method in `GlobalWindController_PG.h` has been commented out and replaced with testing values.
9. The `HashmapTemplate_PG.h` file now includes `VectorMath.h` and uses its functions to calculate distances and positions.
10. The `ShipCannon_PG.h` file now includes `VectorMath.h` and the `normalizeAngle` method has been removed.
11. The `ShipMovementHandler_PG.h` file now includes `VectorMath.h` and the `normalize` and `dot` methods have been removed.
12. The `ShipSailHandler_PG.h` file now includes `VectorMath.h` and a new method `getSailDirectionVector` has been added.
13. The `TextureResourceManager_PG.h` file has had debug print statements removed.
14. The `CollisionManager_PG.cpp`, `EnemyShipHandler_PG.cpp`, `LandMassHandler_PG.cpp`, `Minimap_PG.cpp`, `PlayerShipMovementHandler_PG.cpp`, `ShipCannon_PG.cpp`, and `EnemyShipMovementHandler_PG.cpp` files have replaced distance and position calculations with functions from `VectorMath.h`.
15. The `move` method in `EnemyShip_PG.cpp` and `PlayerShip_PG.cpp` now takes an additional parameter from the `getSailDirectionVector` method in `ShipSailHandler_PG.h`.
16. A new `VectorMath.h` file has been added, which includes a `vm` namespace with several vector math functions.

+++ PirateGame/Header/GlobalMap_PG.h
@@ -104,7 +104,7 @@ namespace PirateGame {
 			if (chunks.find(chunkCoord) == chunks.end()) {
 				// Generate the chunk
 				auto newChunk = std::make_unique<Map>();
-				newChunk->grid(chunkSize.x, chunkSize.y, cellSize, sf::Vector2f(chunkCoord.first * chunkSize.x, chunkCoord.second * chunkSize.y));
+				newChunk->grid(static_cast<int>(chunkSize.x), static_cast<int>(chunkSize.y), cellSize, sf::Vector2f(chunkCoord.first * chunkSize.x, chunkCoord.second * chunkSize.y));
 				chunks[chunkCoord] = std::move(newChunk);
 			}
 		}
